1.12c
Partial support update. Ideal plan: Will continue with 1.12: first until mutli- batchTransaction  a is built, in the sense that all three types are supported. .then until multi- general Batach and jarTransactions are supported. Once all 3 types and multiple of each type is supported, we'll continue on to 1.13 for blockchain101 and/or collapsible dataTables for batch and jarTransactions updates, the both of which will be completed by 1.14.